Extract automation step definitions out into shared-core.
