+++ packages/server/src/automations/actions.ts
@@ -1,3 +1,4 @@
+import { automations } from "@budibase/shared-core"
 import * as sendSmtpEmail from "./steps/sendSmtpEmail"
 import * as createRow from "./steps/createRow"
 import * as updateRow from "./steps/updateRow"
@@ -14,9 +15,7 @@ import * as make from "./steps/make"
 import * as filter from "./steps/filter"
 import * as delay from "./steps/delay"
 import * as queryRow from "./steps/queryRows"
-import * as loop from "./steps/loop"
 import * as collect from "./steps/collect"
-import * as branch from "./steps/branch"
 import * as triggerAutomationRun from "./steps/triggerAutomationRun"
 import * as openai from "./steps/openai"
 import env from "../environment"
@@ -62,27 +61,27 @@ export const BUILTIN_ACTION_DEFINITIONS: Record<
   string,
   AutomationStepDefinition
 > = {
-  SEND_EMAIL_SMTP: sendSmtpEmail.definition,
-  CREATE_ROW: createRow.definition,
-  UPDATE_ROW: updateRow.definition,
-  DELETE_ROW: deleteRow.definition,
-  OUTGOING_WEBHOOK: outgoingWebhook.definition,
-  EXECUTE_SCRIPT: executeScript.definition,
-  EXECUTE_QUERY: executeQuery.definition,
-  SERVER_LOG: serverLog.definition,
-  DELAY: delay.definition,
-  FILTER: filter.definition,
-  QUERY_ROWS: queryRow.definition,
-  LOOP: loop.definition,
-  COLLECT: collect.definition,
-  TRIGGER_AUTOMATION_RUN: triggerAutomationRun.definition,
-  BRANCH: branch.definition,
+  SEND_EMAIL_SMTP: automations.steps.sendSmtpEmail.definition,
+  CREATE_ROW: automations.steps.createRow.definition,
+  UPDATE_ROW: automations.steps.updateRow.definition,
+  DELETE_ROW: automations.steps.deleteRow.definition,
+  OUTGOING_WEBHOOK: automations.steps.outgoingWebhook.definition,
+  EXECUTE_SCRIPT: automations.steps.executeScript.definition,
+  EXECUTE_QUERY: automations.steps.executeQuery.definition,
+  SERVER_LOG: automations.steps.serverLog.definition,
+  DELAY: automations.steps.delay.definition,
+  FILTER: automations.steps.filter.definition,
+  QUERY_ROWS: automations.steps.queryRows.definition,
+  LOOP: automations.steps.loop.definition,
+  COLLECT: automations.steps.collect.definition,
+  TRIGGER_AUTOMATION_RUN: automations.steps.triggerAutomationRun.definition,
+  BRANCH: automations.steps.branch.definition,
   // these used to be lowercase step IDs, maintain for backwards compat
-  discord: discord.definition,
-  slack: slack.definition,
-  zapier: zapier.definition,
-  integromat: make.definition,
-  n8n: n8n.definition,
+  discord: automations.steps.discord.definition,
+  slack: automations.steps.slack.definition,
+  zapier: automations.steps.zapier.definition,
+  integromat: automations.steps.make.definition,
+  n8n: automations.steps.n8n.definition,
 }
 
 // don't add the bash script/definitions unless in self host
@@ -97,7 +96,7 @@ if (env.SELF_HOSTED) {
   BUILTIN_ACTION_DEFINITIONS["EXECUTE_BASH"] = bash.definition
 
   if (env.isTest()) {
-    BUILTIN_ACTION_DEFINITIONS["OPENAI"] = openai.definition
+    BUILTIN_ACTION_DEFINITIONS["OPENAI"] = automations.steps.openai.definition
   }
 }
 
@@ -105,7 +104,7 @@ export async function getActionDefinitions(): Promise<
   Record<keyof typeof AutomationActionStepId, AutomationStepDefinition>
 > {
   if (env.SELF_HOSTED) {
-    BUILTIN_ACTION_DEFINITIONS["OPENAI"] = openai.definition
+    BUILTIN_ACTION_DEFINITIONS["OPENAI"] = automations.steps.openai.definition
   }
 
   const actionDefinitions = BUILTIN_ACTION_DEFINITIONS
