+++ packages/shared-core/src/automations/index.ts
@@ -1,0 +1,1 @@
+export * as steps from "./steps/index"
