+++ packages/shared-core/src/automations/steps/executeScript.ts
@@ -1,0 +1,47 @@
+import {
+  AutomationActionStepId,
+  AutomationCustomIOType,
+  AutomationFeature,
+  AutomationIOType,
+  AutomationStepDefinition,
+  AutomationStepType,
+} from "@budibase/types"
+
+export const definition: AutomationStepDefinition = {
+  name: "JS Scripting",
+  tagline: "Execute JavaScript Code",
+  icon: "Code",
+  description: "Run a piece of JavaScript code in your automation",
+  type: AutomationStepType.ACTION,
+  internal: true,
+  stepId: AutomationActionStepId.EXECUTE_SCRIPT,
+  inputs: {},
+  features: {
+    [AutomationFeature.LOOPING]: true,
+  },
+  schema: {
+    inputs: {
+      properties: {
+        code: {
+          type: AutomationIOType.STRING,
+          customType: AutomationCustomIOType.CODE,
+          title: "Code",
+        },
+      },
+      required: ["code"],
+    },
+    outputs: {
+      properties: {
+        value: {
+          type: AutomationIOType.STRING,
+          description: "The result of the return statement",
+        },
+        success: {
+          type: AutomationIOType.BOOLEAN,
+          description: "Whether the action was successful",
+        },
+      },
+      required: ["success"],
+    },
+  },
+}
