+++ packages/shared-core/src/automations/steps/loop.ts
