+++ packages/shared-core/src/automations/steps/updateRow.ts
@@ -1,0 +1,68 @@
+import {
+  AutomationActionStepId,
+  AutomationCustomIOType,
+  AutomationFeature,
+  AutomationIOType,
+  AutomationStepDefinition,
+  AutomationStepType,
+} from "@budibase/types"
+
+export const definition: AutomationStepDefinition = {
+  name: "Update Row",
+  tagline: "Update a {{inputs.enriched.table.name}} row",
+  icon: "Refresh",
+  description: "Update a row in your database",
+  type: AutomationStepType.ACTION,
+  internal: true,
+  features: {
+    [AutomationFeature.LOOPING]: true,
+  },
+  stepId: AutomationActionStepId.UPDATE_ROW,
+  inputs: {},
+  schema: {
+    inputs: {
+      properties: {
+        meta: {
+          type: AutomationIOType.OBJECT,
+          title: "Field settings",
+        },
+        row: {
+          type: AutomationIOType.OBJECT,
+          customType: AutomationCustomIOType.ROW,
+          title: "Table",
+        },
+        rowId: {
+          type: AutomationIOType.STRING,
+          title: "Row ID",
+        },
+      },
+      required: ["row", "rowId"],
+    },
+    outputs: {
+      properties: {
+        row: {
+          type: AutomationIOType.OBJECT,
+          customType: AutomationCustomIOType.ROW,
+          description: "The updated row",
+        },
+        response: {
+          type: AutomationIOType.OBJECT,
+          description: "The response from the table",
+        },
+        success: {
+          type: AutomationIOType.BOOLEAN,
+          description: "Whether the action was successful",
+        },
+        id: {
+          type: AutomationIOType.STRING,
+          description: "The identifier of the updated row",
+        },
+        revision: {
+          type: AutomationIOType.STRING,
+          description: "The revision of the updated row",
+        },
+      },
+      required: ["success", "id", "revision"],
+    },
+  },
+}
