+++ packages/shared-core/src/index.ts
@@ -5,3 +5,4 @@ export * as utils from "./utils"
 export * as sdk from "./sdk"
 export * from "./table"
 export * from "./themes"
+export * as automations from "./automations"
